// components/input.d.ts
import * as React from 'react';

export interface InputProps {
	'aria-activedescendant'?: string;
	'aria-autocomplete'?: string;
	'aria-controls'?: string;
	'aria-describedby'?: string;
	'aria-expanded'?: boolean;
	'aria-haspopup'?: boolean;
	'aria-labelledby'?: string;
	'aria-owns'?: string;
	'aria-required'?: boolean;
	assistiveText?: {
		label?: string;
		spinner?: string;
	};
	autoComplete?: string;
	children?: React.ReactNode;
	className?: string | string[] | Record<string, boolean>;
	defaultValue?: number | string;
	disabled?: boolean;
	errorText?: React.ReactNode | string;
	fieldLevelHelpTooltip?: React.ReactNode;
	fixedTextLeft?: React.ReactNode | string;
	fixedTextRight?: React.ReactNode | string;
	hasSpinner?: boolean;
	iconLeft?: React.ReactNode;
	iconRight?: React.ReactNode;
	id?: string;
	inlineHelpText?: React.ReactNode | string;
	inputRef?: (ref: HTMLInputElement | null) => void;
	isStatic?: boolean;
	label?: string;
	onBlur?: (event: React.FocusEvent) => void;
	onChange?: (event: React.ChangeEvent<HTMLInputElement>, data?: { value: string | number }) => void;
	onClick?: (event: React.MouseEvent) => void;
	onFocus?: (event: React.FocusEvent) => void;
	onInput?: (event: React.FormEvent) => void;
	onInvalid?: (event: React.FormEvent) => void;
	onKeyDown?: (event: React.KeyboardEvent) => void;
	onKeyPress?: (event: React.KeyboardEvent) => void;
	onKeyUp?: (event: React.KeyboardEvent) => void;
	onSelect?: (event: React.SyntheticEvent) => void;
	onSubmit?: (event: React.FormEvent) => void;
	placeholder?: string;
	minLength?: string;
	minValue?: number;
	maxLength?: string;
	maxValue?: number;
	name?: string;
	readOnly?: boolean;
	required?: boolean;
	role?: string;
	step?: number;
	styleInput?: React.CSSProperties;
	styleContainer?: React.CSSProperties;
	type?: 'text' | 'password' | 'datetime' | 'datetime-local' | 'date' | 'month' | 'time' | 'week' | 'number' | 'email' | 'url' | 'search' | 'tel' | 'color';
	value?: number | string;
	variant?: 'base' | 'counter';
}

declare const Input: React.FC<InputProps>;
export default Input;